types: Import ThreadDimension (formerly ThreadId) from passes
