+++ devito/types/dimension.py
@@ -13,7 +13,7 @@ from devito.types.basic import Symbol, DataSymbol, Scalar
 
 __all__ = ['Dimension', 'SpaceDimension', 'TimeDimension', 'DefaultDimension',
            'SteppingDimension', 'SubDimension', 'ConditionalDimension', 'dimensions',
-           'ModuloDimension', 'IncrDimension', 'ShiftedDimension']
+           'ModuloDimension', 'IncrDimension', 'ShiftedDimension', 'ThreadDimension']
 
 
 Thickness = namedtuple('Thickness', 'left right')
@@ -103,6 +103,7 @@ class Dimension(ArgProvider):
     is_Modulo = False
     is_Incr = False
     is_Shifted = False
+    is_Thread = False
 
     _C_typename = 'const %s' % dtype_to_cstr(np.int32)
     _C_typedata = _C_typename
@@ -1058,6 +1059,34 @@ class ShiftedDimension(IncrDimension):
 
     _pickle_args = ['parent', 'name']
     _pickle_kwargs = []
+
+
+class ThreadDimension(BasicDimension):
+
+    """
+    A special Dimension over threads in a shared-memory parallel program.
+    """
+
+    is_Thread = True
+
+    def __new__(cls, *args, **kwargs):
+        return super().__new__(cls, name='tid', **kwargs)
+
+    def __init_finalize__(self, name, nthreads=None):
+        self._nthreads = nthreads
+
+    @property
+    def symbolic_size(self):
+        return self._nthreads
+
+    def _arg_defaults(self):
+        return {}
+
+    def _arg_values(self, *args, **kwargs):
+        return {}
+
+    _pickle_args = []
+    _pickle_kwargs = ['symbolic_size']
 
 
 def dimensions(names):
